updated some Syntax things and comment every non production cot out

+++ easyfile/index.js
@@ -1,5 +1,5 @@
 
-
+//https://nodejs.org/api/fs.html#filehandlereadoptions
 
 //Include path libary
 var path = require('path');
@@ -36,7 +36,15 @@ function write(dirname, fileName, fileContent) {
 };
 
 //Delete file
+/*TODO: Not Production ready!
 function deleteFile(dirname, fileName) {
+    //look if file exists
+    if(exists(dirname, fileName)){
+        console.error("File don't exist in this directory!");
+        var output = "File don't exist in this directory!";
+        return output;
+        //break funcion
+    }
     fs.unlink(path.join(dirname, fileName), function(err) {
         if (err) {
             return console.error(err);
@@ -45,8 +53,10 @@ function deleteFile(dirname, fileName) {
         return output;
     })
 };
+*/
 
 //read file for a specific word
+/* TODO: Not Production ready!
 function readFor(dirname, fileName, searchedContent) {
     fs.readFile(path.join(dirname, fileName), 'utf8', function(err, data) {
         if (err) {
@@ -56,6 +66,7 @@ function readFor(dirname, fileName, searchedContent) {
         return output;
     })
 };
+*/
 
 //look if file exists
 function exists(dirname, fileName) {
@@ -84,7 +95,7 @@ function exists(dirname, fileName) {
 }
 
 //copy a file from dirname to targetDirname
-/*
+/*TODO: Not Production ready!
 function copy(dirname, fileName, targetDirname, newFileName) {
     fs.copyFile(path.join(dirname, fileName), path.join(targetDirname, newFileName), function(err) {
         if (err) {
@@ -98,6 +109,13 @@ function copy(dirname, fileName, targetDirname, newFileName) {
 
 //rename a file
 function rename(dirname, fileName, newFileName) {
+    //look if file exists
+        if(exists(dirname, newFileName)){
+            console.error("File already exist in this directory!");
+            var output = "File already exist in this directory!";
+            return output;
+            //break funcion
+        }
     fs.rename(path.join(dirname, fileName), path.join(dirname, newFileName), function(err) {
         if (err) {
             return console.error(err);
@@ -108,6 +126,7 @@ function rename(dirname, fileName, newFileName) {
 }
 
 //get the full data of a file
+/* TODO: Not Production ready!
 function read(dirname, fileName) {
     fs.readFile(path.join(dirname, fileName), 'utf8', function(err, data) {
         if (err) {
@@ -117,15 +136,16 @@ function read(dirname, fileName) {
         return output;
     })
 }
+*/
 
 //export all function as amodule
 module.exports = {
     create: create,
     write: write,
     deleteFile: deleteFile,
-    readFor: readFor,
+    //readFor: readFor,
     exists: exists,
     //copy: copy,
     rename: rename,
-    read: read
+    //read: read
 };
